+++ src/main/java/net/geforcemods/securitycraft/screen/SonicSecuritySystemScreen.java
@@ -59,8 +59,9 @@ public class SonicSecuritySystemScreen extends Screen implements ConnectionAcces
 			if (tickCount >= PLAYBACK_DELAY) {
 				if (currentNote < be.getNumberOfNotes()) {
 					NoteWrapper note = be.getRecordedNotes().get(currentNote++);
-					SoundEvent sound = NoteBlockInstrument.valueOf(note.instrumentName().toUpperCase()).getSoundEvent().get();
-					float pitch = (float) Math.pow(2.0D, (note.noteID() - 12) / 12.0D);
+					NoteBlockInstrument instrument = NoteBlockInstrument.valueOf(note.instrumentName().toUpperCase());
+					SoundEvent sound = instrument.hasCustomSound() && !note.customSoundId().isEmpty() ? SoundEvent.createVariableRangeEvent(new ResourceLocation(note.customSoundId())) : instrument.getSoundEvent().get();
+					float pitch = instrument.isTunable() ? (float) Math.pow(2.0D, (note.noteID() - 12) / 12.0D) : 1.0F;
 
 					tickCount = 0;
 					minecraft.level.playSound(minecraft.player, be.getBlockPos(), sound, SoundSource.RECORDS, 3.0F, pitch);